requirements.txt
requests
pyyaml
psutil
openpyxl